fixed bug in ipv6 At() method
it was returning a pointer to stack allocated in6_addr which is undefined behavior. now we return a stack copy of in6_addr. since its a 16 byte array, it is probably more efficient too.
although we could add a new method that returns a direct pointer to the nested fixed string in ipv6 column but that would require modifying the original library code